Refactor fix_all and CK-12 pipeline helpers
