+++ tests/conftest.py
@@ -1,33 +1,18 @@
-"""Pytest configuration and shared fixtures for test suite.
+"""Global pytest configuration for repository-wide test helpers."""
 
-This file contains shared fixtures and configuration that applies across all tests.
-"""
+from __future__ import annotations
 
-import sys
-from collections.abc import Iterator
+from typing import TYPE_CHECKING
 
-import pytest
+if TYPE_CHECKING:
+    import pytest
 
 
-@pytest.fixture(scope="session", autouse=True)
-def restore_pandas_module() -> Iterator[None]:
-    """Ensure pandas module is restored after tests that mock it globally.
-
-    Some tests (like test_query_builder_ui.py) mock pandas in sys.modules
-    to avoid import errors when tkinter isn't available. This fixture ensures
-    the real pandas module is restored so other tests can use it normally.
-
-    Yields:
-        None
-    """
-    # Save the original pandas module if it exists
-    original_pandas = sys.modules.get("pandas", None)
-
-    yield
-
-    # After all tests, restore the original pandas module
-    if original_pandas is not None:
-        sys.modules["pandas"] = original_pandas
-    elif "pandas" in sys.modules:
-        # If pandas was mocked but wasn't originally imported, remove the mock
-        del sys.modules["pandas"]
+def pytest_configure(config: pytest.Config) -> None:
+    """Normalize legacy -k expressions that use pipe separators."""
+    keyword_expr = config.option.keyword
+    if not keyword_expr or "|" not in keyword_expr:
+        return
+    config.option.keyword = " or ".join(
+        part.strip() for part in keyword_expr.split("|")
+    )
